feat: Adding support for Tilesets referencing files
In Tiled, a tileset can reference a .tsx or .tsj file.
Previously, those "non embedded" tilesets where not properly parsed, though valid.

This new version validates those tilesets.

In order to discriminate between the 2 typeset types, you will now have to check if the "source" or "name" field is set.

```
if ("name" in tileset) {
    // tileset is an embedded tileset
}

if ("source" in tileset) {
    // tileset is an reference to an external tileset
}
```